добавление комментариев

+++ books/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-from .models import Book, Genre, Publisher, Tag
+from .models import Book, Genre, Publisher, Tag, Comment
 from django.http import HttpResponse
 from .forms import BookForm
 
@@ -35,38 +35,41 @@ def get_tag_books(request, title):
     return render(request, "tag_detail.html", context={"tag_books": tag_books})
 
 def add_book(request):
-    if request.method == "GET":
-        form = BookForm()
-        return render(request, "add_book.html", context={"form": form})
-    elif request.method == "POST":
-        publisher_id = request.POST['publisher']
-        genre_id = request.POST['genre']
+    if request.user.is_authenticated:
+
+        if request.method == "GET":
+            form = BookForm()
+            return render(request, "add_book.html", context={"form": form})
+        elif request.method == "POST":
+            publisher_id = request.POST['publisher']
+            genre_id = request.POST['genre']
 
 
-        publisher = None
-        genre = None
-        image = request.FILES.get('image', "default.jpg")
+            publisher = None
+            genre = None
+            image = request.FILES.get('image', "default.jpg")
 
+            if publisher_id != '':
+                publisher = Publisher.objects.get(id=publisher_id)
 
-        if publisher_id != '':
-            publisher = Publisher.objects.get(id=publisher_id)
+            if genre_id != '':
+                genre = Genre.objects.get(id=genre_id)
 
-        if genre_id != '':
-            genre = Genre.objects.get(id=genre_id)
-        print(request.FILES)
+            book = Book.objects.create(title = request.POST['title'],
+                                author = request.POST['author'],
+                                year=request.POST['year'],
+                                raiting=request.POST['raiting'],
+                                publisher=publisher,
+                                genre=genre,
+                                image=image,
+                                user = request.user)
+            tags = request.POST.getlist('tags')
+            book.tags.set(tags)
+            book.save()
 
-        book = Book.objects.create(title = request.POST['title'],
-                            author = request.POST['author'],
-                            year=request.POST['year'],
-                            raiting=request.POST['raiting'],
-                            publisher=publisher,
-                            genre=genre,
-                            image=image)
-        tags = request.POST.getlist('tags')
-        book.tags.set(tags)
-        book.save()
-
-        return redirect("books")
+            return redirect("books")
+    else:
+        return HttpResponse("<h1>У вас нет прав на это действие!</h1>")
 
 def search_book(request):
     title = request.GET['title']
@@ -86,8 +89,11 @@ def delete_book(request, id):
         book = Book.objects.get(id=id)
     except Book.DoesNotExist:
         return HttpResponse(f"<h1> Книги с таким айди: {id} не существует </h1>")
-    book.delete()
-    return redirect('books')
+    if request.user.username != book.user.username:
+        return HttpResponse("<h1> У Вас нет прав на обновление книги! </h1>")
+    else:
+        book.delete()
+        return redirect('books')
 
 def update_book(request, id):
     try:
@@ -95,36 +101,51 @@ def update_book(request, id):
     except Book.DoesNotExist:
         return HttpResponse(f"<h1>Книги с таким айди: {id} не существует!</h1>")
 
-    if request.method == "GET":
-        form = BookForm(instance=book)
+    if request.user.username != book.user.username:
+        return HttpResponse("<h1> У Вас нет прав на обновление книги! </h1>")
 
-        return render(request, "update_book.html", context={"form": form,
-                                                            "book": book})
     else:
-        publisher_id = request.POST['publisher']
-        genre_id = request.POST['genre']
+        if request.method == "GET":
+            form = BookForm(instance=book)
 
-        publisher = None
-        genre = None
-        image = request.FILES.get('image', "default.jpg")
+            return render(request, "update_book.html", context={"form": form,
+                                                                "book": book})
+        else:
+            publisher_id = request.POST['publisher']
+            genre_id = request.POST['genre']
 
-        if publisher_id != '':
-            publisher = Publisher.objects.get(id=publisher_id)
+            publisher = None
+            genre = None
+            image = request.FILES.get('image', "default.jpg")
 
-        if genre_id != '':
-            genre = Genre.objects.get(id=genre_id)
+            if publisher_id != '':
+                publisher = Publisher.objects.get(id=publisher_id)
 
-        book.title = request.POST['title']
-        book.author = request.POST['author']
-        book.year = request.POST['year']
-        book.raiting = request.POST['raiting']
-        book.publisher = publisher
-        book.genre = genre
-        book.image = image
-        tags = request.POST.getlist('tags')
-        book.tags.set(tags)
+            if genre_id != '':
+                genre = Genre.objects.get(id=genre_id)
 
-        book.save()
+            book.title = request.POST['title']
+            book.author = request.POST['author']
+            book.year = request.POST['year']
+            book.raiting = request.POST['raiting']
+            book.publisher = publisher
+            book.genre = genre
+            book.image = image
+            tags = request.POST.getlist('tags')
+            book.tags.set(tags)
 
-        return redirect (get_book, id=book.id)
+            book.save()
+
+            return redirect (get_book, id=book.id)
+
+def add_comment(request, id):
+    print(request.POST)
+    raiting = 5
+    book=Book.objects.get(id=id)
+    Comment.objects.create(content = request.POST['comment'],
+                           raiting = raiting,
+                           user = request.user,
+                           book = book)
+
+    return redirect("get_book", id=id)
